+++ N-Body/QuadParticleTree.cpp
@@ -120,8 +120,8 @@ Particle& QuadParticleTree::get_center_of_mass_particle(Particle& input_particle
 	} else {
 		// Get distances	
 		Particle center_of_mass_particle = Particle(center_of_mass_x_, center_of_mass_y_, total_mass_);
-		double distance_from_center_of_mass = input_particle.get_distance(center_of_mass_particle);
-		double side = get_side_size();
+		float distance_from_center_of_mass = input_particle.get_distance(center_of_mass_particle);
+		float side = get_side_size();
 
 		if (side / distance_from_center_of_mass < THETA) {
 			return center_of_mass_particle;
@@ -142,8 +142,8 @@ void QuadParticleTree::apply_acceleration(Particle& input_particle) const {
 	} else {
 		// Get distances	
 		Particle center_of_mass_particle = Particle(center_of_mass_x_, center_of_mass_y_, total_mass_);
-		double distance_from_center_of_mass = input_particle.get_distance(center_of_mass_particle);
-		double side = get_side_size();
+		float distance_from_center_of_mass = input_particle.get_distance(center_of_mass_particle);
+		float side = get_side_size();
 
 		if (side / distance_from_center_of_mass < THETA) {
 			input_particle.add_acceleration(center_of_mass_particle);
